Case-sensitive walletAddress queries pulls no results
