+++ src/ProductPayments.tsx
@@ -27,33 +27,14 @@ const ProductPayments: React.FC = () => {
   // payment
   const [paymentAmount, setPaymentAmount] = useState(0);
 
-  const testOutstandingNftBalance: OutstandingNftBalance = {
-    balanceRemaining: 1.0,
-    balanceStart: 2.0,
-    buyerAddress: "0x52554BfE4baC4aE605Af27A2e131480F2D219Fe6",
-    createdAt: 1649908881,
-    id: "00000",
-    nftContractAddress: "0x08207fe7f1f7c9f1c39e4720b9f7bfe2afd01907",
-    nftTokenId: 11,
-    product: {
-      description: 'A new revolutionary product A new revolutionary product A new revolutionary product A new revolutionary product A new revolutionary product A new revolutionary product A new revolutionary product',
-      productImageUrls: ['https://images-na.ssl-images-amazon.com/images/I/91TvWl33h4L.jpg', "https://i.kym-cdn.com/entries/icons/mobile/000/006/026/NOTSUREIF.jpg", "https://upload.wikimedia.org/wikipedia/commons/thumb/e/e0/Giraffa_camelopardalis_angolensis.jpg/1024px-Giraffa_camelopardalis_angolensis.jpg"],
-      id: "00000",
-      isListed: true,
-      name: 'Guide to the Universe Book',
-      price: 10000,
-      quantity: 100,
-    },
-    sellerAddress: "0x2929C3c9805dD1A16546251b9b0B65583FD302c8"
-  }
-
   // First function called in useEffect
   const connectWalletAndQueryBalances = async () => {
-    console.log(`Connecting wallet and querying balances`);
+
     // Connect Wallet if needed
     const resp:any = await connectWallet();
     const addr = resp.address;
     console.log(`ADDRESSS ${addr}`);
+
     if (addr !== "") {
       accountChangedHandler(addr);
     } else {
@@ -65,6 +46,8 @@ const ProductPayments: React.FC = () => {
 
     setWalletAddress(newAccount);
     getUserBalance(newAccount.toString());
+
+    console.log(`GONNA TEST CASES: ${newAccount} /// VS /// ${newAccount.toString()}`);
 
     if (!callingBalances) {
       console.log(`SHOULD CALLLL`)
@@ -97,11 +80,12 @@ const ProductPayments: React.FC = () => {
 
     const exactAmount = await getExactPaymentleft("0x52554BfE4baC4aE605Af27A2e131480F2D219Fe6", currentOustandingBalance!.nftContractAddress, currentOustandingBalance!.nftTokenId);
 
-    await repayStore(currentOustandingBalance!.nftContractAddress, currentOustandingBalance!.nftTokenId, exactAmount);
+    await repayStore(currentOustandingBalance!.id, currentOustandingBalance!.nftContractAddress, currentOustandingBalance!.nftTokenId, exactAmount);
 
   }
 
   useEffect(() => {
+    console.log(`NOW IN THE PAYMENTS PAGE!`);
     connectWalletAndQueryBalances();
   }, [])
 
@@ -120,8 +104,8 @@ const ProductPayments: React.FC = () => {
               <div className="nft-content">
                 { oustandingBalancesArr.map((balanceObj:OutstandingNftBalance, index:any) => {
                   return (
-                    <a onClick={() => { console.log(JSON.stringify(balanceObj)); selectedOutstandingBalance(balanceObj)}}>
-                      <div className="balance-box" key={index}>
+                    <a onClick={() => { console.log(JSON.stringify(balanceObj)); selectedOutstandingBalance(balanceObj)}} key={index}>
+                      <div className="balance-box">
                         <div className="balance-img-div">
                           <img className="balance-img" src={balanceObj.product.productImageUrls[0]} />
                         </div>
@@ -136,42 +120,6 @@ const ProductPayments: React.FC = () => {
               </div>
               }
             </Row>
-
-            {/* <Row>
-              <div className="balance-box-unselected">
-                <div className="balance-img">
-                  <img />
-                </div>
-                <div className="balance-info">
-                  <p className="balance-title">Universe Book</p>
-                  <p className="balance-amount">Remaining Balance: </p>
-                </div>
-              </div>
-            </Row>
-
-            <Row>
-              <div className="balance-box-unselected">
-                <div className="balance-img">
-                  <img />
-                </div>
-                <div className="balance-info">
-                  <p className="balance-title">Universe Book</p>
-                  <p className="balance-amount">Remaining Balance: </p>
-                </div>
-              </div>
-            </Row>
-
-            <Row>
-              <div className="balance-box-unselected">
-                <div className="balance-img">
-                  <img />
-                </div>
-                <div className="balance-info">
-                  <p className="balance-title">Universe Book</p>
-                  <p className="balance-amount">Remaining Balance: </p>
-                </div>
-              </div>
-            </Row> */}
           </Col>
           <Col xs={12} md={6} className="right-col">
             { oustandingBalancesArr.length <= 0 &&
